requirements.txt
pandas
boto3
streamlit
python-dotenv
snowflake-connector-python
groq
botocore
